+++ docs/versionhistory.rst
@@ -9,6 +9,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - Fixed ``move_on_after()`` and ``fail_after()`` getting stuck on curio in some circumstances
 - Fixed socket operations not allowing timeouts to cancel the task
 - Harmonized the default task names across all backends
+- ``wait_all_tasks_blocked()`` no longer considers tasks waiting on ``sleep(0)`` to be blocked
+  on asyncio and curio
 
 **1.3.1**